requirements.txt
streamlit
paho-mqtt
# Add other dependencies here
